version 1.5.6
The rest of the fix for Issue 5: Popup disappears with it shouldn't.

Added pagehide and hashchange handlers to popdown when going to a different page
(including deviantart's /# URLs).

Changed preferred location of non-overlapping popup to be below, above, left, right;
used to put right first.  Right is the worst since it's the most likely
to be covered by a tooltip (eg on deviantart.com).