Version 1.6.1alpha2.  EXPERIMENTAL VERSION with debug on.
Experimental possible workaround for Linux narrow pop-up
problem.  Instead of closing popup it just moves it to the
corner of the screen and makes it small; it also sets the
size of the popup explicitly instead of relying on setting
CSS on the image.

Debug log is turned on so a particuar user can test it
and send me the log.  NOT FOR GENERAL USE DUE TO PRIVACY
CONCERNS OF THE LOG FILE.