Changed how popup position is set:
Explicitly calculates screen-space x and y of popup, and positions it
that way instead of letting Firefox position it relative to its thumbnail.
This allows us to position it consistently even when doing the Linux
workaround (which doesn't re-pop up the image).

Also implemented better logic than we've ever had for positioning
when overlapping the thumb.  It now centers the popup in that situation
unless it can partially reveal the thumb by offsetting to an edge of the
window.  Previously it always just put it in the upper-left corner.