Fixed #51: Hovering on a magnet link in TPB triggers the link.
The worst symptom was already fixed by a prior change: instead of trying
to open a magnet: link, it was just trying to open a gif link without
any protocol (//static.thepiratebay.se/img/icons/icon-magnet.gif).

The present change fixes the latter problem.  The Fix was to apply the document's
baseURI after getting it from a background style.

I also added another layer of security check to ensure the URL
starts with http or https to prevent a magnet: link from being
used as an image.

Added these issues to TestCases.html.