Fixed issue #56: Google images popup may stay up from prior image
We no longer set an ignore-bbox when popping up the image; instead we
set it only when dismissing the popup due to events which may immediately
cause another mouseOver event, and so where we must prevent an undesired
pop-up; examples are closing by Escape, clicking, or focus loss.

This fixes the problem since when Google's popup disappears, we now
handle the generated mouseOver event and pop down our own popup, and
if we end up over a different thumbnail, initiate a popup for it.

This also half-fixes Issue #19: Popup sometimes doesn't pop down when move out of thumb.
It fixes so moving outside the visible edges of the thumb dismiss the popup
when the popup isn't overlapping the thumb.  But there are still problems
when dismising via Escape, click, or focus loss; the same or a different popup
can't be initiated until the mouse moves outside and back inside the larger
unclipped thumb's extents.  Similarly, if the popup overlaps the thumb, it isn't
dismissed until the mouse moves outside the unclipped thumb's extents.