Fixed issue #57: Popup cycling on Google Visually Similar Images small thumbs
The fix is to stop propagation of the mouseOver event specifically
on the small thumbs of Visually Similar Images in google results,
so Google's popup doesn't appear and they cycling is avoided.

The technique might be useful too as an option on Google Images,
though if we suppressed the popup there we couldn't get a caption
for the Thumbnail Zoom Plus pop-up.