Fixed #60: Firefox doesn't show link URL after popup displays
The url tooltip seemed to disappear as a result of the 'blur' (mouseout) event the
document receives as the popup gets focus. I'm able to work around the problem by
explicitly calling blur() but then sending a synthetic mouseOver event, and then popping up.

This seems to work, but I'm concerned about the risk of the synthetic event
causing problems in firefox or incompatibility with some plug-in, especially
if no corresponding blur() is ever sent later.