Fixed #65: Toolbutton & hotkey to enable/disable TZ
Toolbar is now a button with a small menu arrow to its right.
Preferences have a "Enable" checkbox.
Pressing 'x' while popped-up toggles the Enable preference, and if it's
now disabled, closes the popup.
Even when disabled, popup will appear if "Key to maximize the image"
is down (shift by default).