Fix #44: Large images load in too large chunks Improved #66: gif animation starts before popup displayed For the latter, reduced delay before showing for .gif, but only helps when URL contains ".gif".
NOTE: Fix for #44 changes how images are displayed and reduces _panelWidthAddon
to 0 (based on testing on mac).  Will need testing on all operating
systems and multiple firefox versions.  There is still a slight
problem of an extra line 4 or so pixels high below the image, above
the caption.  That seems to be space in the div outside the image, but
I haven't been able to eliminate it even setting border, margin, and
pad all to 0px.