Bug fix: applies 'disabled' checkboxes more strongly to Others, Thumbnails:
Previously, the Others rule specificallly checked for certain sites including
youtube and tried to disable Others for youtube when youtube was disabled.
But that didn't apply to Thumbnails, and not even to Thumbnails triggered
due to Others being enabled.

Now the rule is applied in general to for all site checkboxes, affecting
both Others and Thumbnails.  If a site's checkbox is off, Others
and Thumbnails will be disallowed on hosts matching the disabled pages.
Note that this tests only against the page containing the thumb/link,
not against the target of the thumb/link.

Problem reported via mozilla add-ons site by "CoffeeCup".