Added workaround for #67: Linux - thumbnails no longer display
New preference "Pop-up Takes Keyboard Focus" defaults to On; setting
it Off restores the behavior of 1.6.0, where the pop-up does not
grab focus and is not dismissed if it loses focus.

The draw-backs are that the user still sees the text cursor when the pop-up
is displayed, and so may start typing into the web page, but typing
will be interpreted as TZP hotkeys.  And the pop-up won't automatically
dismiss if the user does e.g. Control-F to "find"; when the user
starts typing his "find" text, typing is interpreted as TZP hotkeys.