Fixed #62: Saved images are very large
Now directly saves the contents of the image's URL instead of
rendering into a Canvas and saving the PNG version of that.
So the result is the same size and format as the original.
It's also faster since it eliminates the separate step of
downloading and rendering in a canvas.

Caveats:
1. The original format is used, even if the user changes the file extension
in the Save dialog
2. At least on Mac in Firefox 12, when an existing file is clicked in
the Save dialog, Firefox doesn't enter its extension into the Save As
field; if the user doesn't manually add the extension, the resulting
file will have no extension.  That seems to be Firefox's normal behavior;
eg it happens when saving an html page (outsize of TZP).
3. So far tested only on Mac with Firefox 12.

Old behavior is still available by changing saveInPngFormat in downloadImage().