Fixed #77: gif animation restarts when fully loaded
This works around Firefox bug 756367 by using an xul <image>
tag rather than an <html:img> tag in the popup for *.gif images.
Note that the workaround won't take effect if a gif image's
URL does not include the string ".gif".

Caveat: the 'working' icon doesn't appear for .gif images
since it's in the hidden panelHtmlImage.