New feature: [ and ] keys move to prev and next image numerically.
Example pages where it works on some of the links:

http://www.google.com/search?um=1&hl=en&safe=off&tbm=isch&sa=1&q=poodle&oq=poodle&gs_l=img.3..0l10.31700.32746.8.33214.6.4.0.2.2.0.161.529.0j4.4.0...0.0...1c.1Lx1fRodnDE&pbx=1&biw=1056&bih=755&bav=on.2,or.r_gc.r_pw.r_qf.&cad=b&sei=0Z4oUPywOZPlqQHo9oDQCQ&cad=h

http://my.opera.com/Milano1/albums/showpic.dml?album=7138392&picture=107119352

Limitations:

- If the generated image doesn't actually exist, you'll see the error indicator briefly.
Ideally you'd be able to reverse direction and get back to a valid image, but it
doesn't support that.

- When going backward from a number which isn't 0-padded (eg 100)
it can't tell whether it should 0-pad (099) or not (99).  In
that situation it does not zero-pad.  When incrementing it can correctly
deduce appropriate 0-padding.