Now copies image or URL for any type of image.
Tested on various png, jpg, gif on reddit.  Tested on
mac pasting into Word, Photoshop, Thunderbird.

Changes and adds hotkeys:

Toggle caption is now on 'i' as in Info

Copy image contents if 'c'

Copy image location is 'l' as in Location

Issue:

When copying a large jpg image, the bottom 10% or so of the
image appears black and the image size is *much* larger than the
original.  Firefox seems to be converting the jpg data to png,
even though TZP isn't requesting that.

Example:
http://www.reddit.com/tb/yt19r

Using Firefox's built-in right-click Copy Image on
a jpg image does not have these problems.