Cherry-picked from dev branch:
Another (more general) fix for 30-second stall in "Share" popup dialog of
some Zynga games on Facebook, eg The Ville.  Problem was analyzing a
data: URL which contained full image data and so was very large;
parsing regular expressions against this was very slow until the
previous commit, which optimized the regexp's in the Thumbnails
rule.  But as another level of protection, we don't try to show
pop-ups for URLs greater than 1000 chars long.