Restores ability to see image as it's loading.  Fixes problem with prev image briefly appearing at start of next pop-up.
Works around bug starting in ff 19 or 20 where the prior image
still displays in an img, even though its src was previously cleared;
that bug also caused prior image's width to be returned in naturalWidth,
interfering with TZP's ability to know when the image is loaded
enough to start dispaying it.

Works around it by destroying and recreating the img node
when the pop-up is closed.  Seems to work but needs more testing
and verification of no memory leak.

Version 2.5alpha1.