imgur.com fix for imgur.com/gallery/ links.
It used to assume the link was to a single-image gallery, and
directly display the corresponding jpg.  But occasionally the link is
to a gallery of an album, and the generated link is to an unrelated image.

To avoid that problem, we display /gallery/ links using
OthersIndirect, which gets the right result in both cases,
but is slower.