Fixed #118: Site requests: furaffinity.net
Now uses Others (Indirect) to read the linked 'View' page
to get the high-rez URL.  So it displays a higher-resolution
pop-up than before.

Caveats:
1. If the image author marked the Download link for logged-in
users only, you'll need to be logged in.
2. On an actual View or Submit page hover the Downlaod link rather
than the photo to see higher resolution (or click the
photo -- a feature of the site itself).

Fix will be in 3.0beta6 and 3.0.