Right-click context menu no longer affected unless RMB is configured.
This is a risk reduction strategy, so that TZP doesn't give
right-clicks any special behavior unless Right Mouse Button is configured
in its preference.

This is a partial workaround for right-click not showing
spell correction suggestions.  The problem no longer happens
except when Right Mouse Button is set in TZP preferences.