Fixed #177: Right mouse button option in 3.2beta1 prevents spell check corrections.
The fix is to allow the context menu to pop up immediately on text areas, which allows the
spell checker to show suggestions.  This means if you try to right-drag from a text area onto a
link or image, you'll see a context menu, but that's probably not a common usage.