Fixed #185: on imgur.com must hover center of thumb to get full pop-up
Two fixes:
1. added logic to detect bottom and top third, and adjust what node it
uses so it still finds the correct img node.

2. changed Others Indirect to remove "?fb" from URL, as suggested
by Hakkin.

Note: with these fixes, the specific problem raised in this
issue is fixed but a problem remains: when hovering
a thumb for a multi-image gallery on imgur.com,
it shows just a single image (unless you hover one of
the textual links).