refines previous fix for #185: on imgur.com must hover center of thumb to get full pop-up
Now forces it not to use the Imgur or Thumbnail rule if the thumbnail is
linked to an imgur gallery.  Instead, it'll use the Others (Indirect) rule
(if enabled) so it can detect if it's a multi-image gallery, and handle it
appropriately.  Fixes that it wasn't showing as gallery, though it'll be
a little slower.