Fix #178: High CPU/GPU usage & fan running when using TZP in Firefox 33.0+ on Windows
The underlying firefox bug is still present (in ff 35.0.1) but we
work around it by avoiding the trigger.  We no longer
try to position a pop-up where it could go beyond the bottom or
right edge of the available screen space (which excludes the task bar).
So the pop-up doesn't overlap the Windows task bar and the
bug is avoided.

Pop-up positioning also looks better to the user: instead of
showing a larger pop-up but seeing only the top half of it,
the user sees the entire pop-up, though at smaller size.

Verified it fixes the problem in Windows 10 Technical Preview build 9926.