Bug fix: toolbar button in menu pane wasn't showing correct state.
To repro: disable TZP.  Make sure its icon is in the menu-button pane.

Open new window.  Open menu pane.  Notice that the TZP icon appears enabled.

This is now fixed by having the button make a callback to
set the right state when it is displayed.  Apparently the
button was being created later than the main TZP window so the
original trigger wasn't working.