Disabled showing imgur.com *.gifv since doesn't work.
In June 2016 imgur.com changed their site so that gifv images
are no longer webm's, and won't display in a Firefox add-on
pop-up like TZP anymore.  To avoid just getting an error icon, we
no longer try to show pop-ups for gifv's.

An alternative would be to show them as .gif's, but they'd be
lower quality and use more bandwidth than if the user manually opened
the link.  Uncommenting the .gifv line would achieve that.

Related to #238.