Fixed #245: instagram not working for some users
For some users, instagram serves images from instagram.fyto1-1.fna.fbcdn.net
rather than scontent-sjc2-1.cdninstagram.com.  This host wasn't working, but
does now.

Problem analysis:

It wasn't working because the rule in TZP disallowed "?fb" from the URL
since on imgur.com that had special undesired meaning.  The "?" was supposed
to be backslash-quoted to treat the question mark as a literal.
But the expression used a double backslash (\\), which is appropriate
in a double-quoted string, but in a /.../ expression like this
it needs to use a single backslash.  So the "?" was treated as a special
character, and "fb" appearing anywhere in the URL was disallowed,
preventing fbcdn.net from working.